+++ zscaler/zcon/client.go
@@ -13,6 +13,7 @@ import (
 
 	"github.com/SecurityGeekIO/zscaler-sdk-go/v3/cache"
 	"github.com/SecurityGeekIO/zscaler-sdk-go/v3/logger"
+	"github.com/SecurityGeekIO/zscaler-sdk-go/v3/zscaler/errorx"
 	"github.com/google/uuid"
 )
 
@@ -97,7 +98,7 @@ func (c *Client) Request(endpoint, method string, data []byte, contentType strin
 
 		resp.Body.Close()
 		if resp.StatusCode > 299 && resp.StatusCode != http.StatusUnauthorized {
-			return nil, checkErrorInResponse(resp, fmt.Errorf("api responded with code: %d", resp.StatusCode))
+			return nil, errorx.CheckErrorInResponse(resp, fmt.Errorf("api responded with code: %d", resp.StatusCode))
 		}
 	}
