+++ zscaler/zdx/client.go
@@ -17,6 +17,7 @@ import (
 
 	"github.com/SecurityGeekIO/zscaler-sdk-go/v3/logger"
 	"github.com/SecurityGeekIO/zscaler-sdk-go/v3/utils"
+	"github.com/SecurityGeekIO/zscaler-sdk-go/v3/zscaler/errorx"
 	"github.com/google/go-querystring/query"
 	"github.com/google/uuid"
 )
@@ -219,7 +220,7 @@ func (client *Client) do(req *http.Request, v interface{}, start time.Time, reqI
 	logger.LogResponse(client.Config.Logger, resp, start, reqID)
 	logger.WriteLog(client.Config.Logger, "Response Body: %s", string(respBody)) // Log the response body separately
 
-	if err := checkErrorInResponse(resp); err != nil {
+	if err := errorx.CheckErrorInResponse(resp, err); err != nil {
 		return resp, err
 	}
